requirements.txt
flask==2.0.3
google-cloud-storage
google-cloud-datastore==2.7.1
